Extending the example with classes. TypeScript supports new features in JavaScript, like support for class-based object-oriented programming.
Here we’re creating a Student class with a constructor and a few public fields. Notice that classes and interfaces play well together, letting the programmer decide on the right level of abstraction.

Also of note, the use of public on arguments to the constructor is a shorthand that allows us to automatically create properties with that name.